+++ tests/terminalPaths.test.js
@@ -11,19 +11,19 @@ test('normalizePath: removes ".", collapses slashes, handles ".."', () => {
 });
 
 test('resolvePath: empty -> currentDir', () => {
-  assert.equal(resolvePath('', { currentDir: '/user/ryan', homeDir: '/user/ryan' }), '/user/ryan');
-  assert.equal(resolvePath(null, { currentDir: '/user/ryan', homeDir: '/user/ryan' }), '/user/ryan');
+  assert.equal(resolvePath('', { currentDir: '/home/rg', homeDir: '/home/rg' }), '/home/rg');
+  assert.equal(resolvePath(null, { currentDir: '/home/rg', homeDir: '/home/rg' }), '/home/rg');
 });
 
 test('resolvePath: ~ expansion', () => {
-  assert.equal(resolvePath('~', { currentDir: '/tmp', homeDir: '/user/ryan' }), '/user/ryan');
-  assert.equal(resolvePath('~/bin', { currentDir: '/tmp', homeDir: '/user/ryan' }), '/user/ryan/bin');
+  assert.equal(resolvePath('~', { currentDir: '/tmp', homeDir: '/home/rg' }), '/home/rg');
+  assert.equal(resolvePath('~/bin', { currentDir: '/tmp', homeDir: '/home/rg' }), '/home/rg/bin');
 });
 
 test('resolvePath: absolute and relative', () => {
   assert.equal(resolvePath('/etc/../usr', { currentDir: '/x', homeDir: '/h' }), '/usr');
-  assert.equal(resolvePath('Documents', { currentDir: '/user/ryan', homeDir: '/user/ryan' }), '/user/ryan/Documents');
-  assert.equal(resolvePath('../Documents', { currentDir: '/user/ryan/bin', homeDir: '/user/ryan' }), '/user/ryan/Documents');
+  assert.equal(resolvePath('Documents', { currentDir: '/home/rg', homeDir: '/home/rg' }), '/home/rg/Documents');
+  assert.equal(resolvePath('../Documents', { currentDir: '/home/rg/bin', homeDir: '/home/rg' }), '/home/rg/Documents');
 });